+++ app/models/checkbox_answers.rb
@@ -2,7 +2,9 @@ class CheckboxAnswers < ActiveRecord::Base
   self.implicit_order_column = "created_at"
   belongs_to :step
 
-  validates :response, presence: true
+  validates :response,
+    presence: true,
+    unless: proc { |answer| answer.step.skippable? && answer.skipped }
 
   def response=(args)
     return if args.blank?