+++ db/migrate/20210301151653_add_skippable_to_step.rb
@@ -1,0 +1,5 @@
+class AddSkippableToStep < ActiveRecord::Migration[6.1]
+  def change
+    add_column :steps, :skip_call_to_action_text, :string, default: nil
+  end
+end