+++ db/migrate/20210302110746_add_skipped_to_checkbox_answers.rb
@@ -1,0 +1,5 @@
+class AddSkippedToCheckboxAnswers < ActiveRecord::Migration[6.1]
+  def change
+    add_column :checkbox_answers, :skipped, :boolean, default: false
+  end
+end