+++ spec/features/visitors/anyone_can_complete_a_journey_spec.rb
@@ -302,6 +302,28 @@ feature "Anyone can start a journey" do
     end
   end
 
+  context "when the question is skippable" do
+    scenario "allows the user to not select an answer" do
+      start_journey_from_category_and_go_to_question(category: "skippable-checkboxes-question.json")
+
+      click_on(I18n.t("generic.button.next"))
+      expect(page).to have_content("can't be blank")
+
+      check("Lunch")
+      click_on("None of the above")
+
+      within(".app-task-list") do
+        expect(page).to have_content("Complete")
+      end
+
+      click_first_link_in_task_list
+
+      expect(page).not_to have_checked_field("Lunch")
+
+      expect(CheckboxAnswers.last.skipped).to be true
+    end
+  end
+
   context "when the Contentful model is of type staticContent" do
     context "when Contentful entry is of type paragraphs" do
       scenario "the content is not displayed in the task list" do