+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - radio questions can be configured with an "or" separator
 - specification can show further information for radio answers
 - specification can show further information for checkbox answers
+- checkbox answers can be completed without choosing a given answer
 
 ## [release-005] - 2021-1-19